Fix the UID cache might not be synchronized when copying directories externally
Validate cached entries for uids and purge invalid entries instead
of purging all entries, which keeps the uids of files as unchanged
as possible.

Automatically assign new UIDs to duplicate new files. This prevents
the file path corresponding to the UID in the UID cache from being
replaced (from the perspective of the file path, the UID of the file
has changed).