Prioritize directory/file removal actions
During `scan_fs_changes()`, an insertion point is added to `scan_actions`,
and directory/file removal actions are inserted after this insertion point,
so that directory/file removal actions are processed first.

When used in conjunction with `ResourceUID`, it will provide a more reliable
basis for file movement tracking.

During the same `scan_fs_changes()` call, if a `uid` is **first** deleted
from `ResourceUID` due to a file being removed, and **then** added again to
`ResourceUID` due to a file being added, we can consider this a **file move**
operation.