Subdivide the focus state of the window on X11.
1. Use `focused` to track `_NET_WM_STATE_FOCUSED`;
2. Use `contents_focused` to track `FocusIn`/`FocusOut` events.

The first point prevents windows/programs from losing focus when dragging
or resizing the window or performing certain time-consuming tasks. Windows
set to `no_focus` will be failed to have their `focused` property set to
`true`.

The `focused` and `no_focus` properties could previously both be `true`
simultaneously, so the second point is for compatibility (to avoid not
receiving the release event).